+++ next.config.js
@@ -1,7 +1,8 @@
 module.exports = {
   images: {
     domains: ['cdn.sanity.io'],
-    deviceSizes: [412, 524, 632, 731, 812, 1024, 1640, 1920],
+    deviceSizes: [439, 632, 756, 828, 1080, 1512],
+    imageSizes: [],
   },
   webpack: (config, { dev, isServer }) => {
     if (isServer) {
